snapshot: before one-click cleanup
