+++ src/actions/__init__.py
@@ -1,2 +1,1 @@
-# Compat shim: from actions import X → smart_mail_agent.actions.X
 from smart_mail_agent.actions import *  # noqa: F401,F403
