+++ src/policy_engine.py
@@ -1,79 +1,2 @@
-from __future__ import annotations
-
-import os
-from typing import Any, Dict, Iterable, Optional
-
-import yaml
-
-
-def _sum_attachments_size(att: Optional[Iterable[dict]]) -> int:
-    total = 0
-    for a in att or []:
-        try:
-            total += int(a.get("size") or 0)
-        except Exception:
-            pass
-    return total
-
-
-def _from_domain(addr: Optional[str]) -> Optional[str]:
-    if not addr or "@" not in addr:
-        return None
-    return addr.split("@", 1)[1].lower()
-
-
-def _detect_roles(a: Dict[str, Any], b: Dict[str, Any]) -> tuple[Dict[str, Any], Dict[str, Any]]:
-    """回傳 (result, request)；自動判別參數順序以相容舊測試。"""
-    score_a = int(bool(a.get("action_name") or a.get("ok") or a.get("code")))
-    score_b = int(bool(b.get("action_name") or b.get("ok") or b.get("code")))
-    if score_a > score_b:
-        return a, b
-    if score_b > score_a:
-        return b, a
-    # 平手時用特徵判斷：含 predicted_label/attachments 視為 request
-    if "predicted_label" in a or "attachments" in a:
-        return b, a
-    return a, b
-
-
-def apply_policies(x: Dict[str, Any], y: Dict[str, Any], policy_path: str = "config/policy.yaml") -> Dict[str, Any]:
-    """
-    低信心簽審（預設閾值 0.6；可在 YAML low_confidence_review.threshold 覆蓋）
-    - 若低於閾值：result.meta.require_review=True，並合併 cc。
-    - 相容舊參數順序：自動判別 (result, request)。
-    """
-    result, request = _detect_roles(x, y)
-    res = dict(result or {})
-    meta = dict(res.get("meta") or {})
-    cc = list(res.get("cc") or [])
-
-    conf = request.get("confidence")
-    threshold = 0.6
-    extra_cc = ["review@company.com"]  # 預設 cc（測試期望至少包含此位址）
-
-    try:
-        if os.path.exists(policy_path):
-            rules = yaml.safe_load(open(policy_path, "r", encoding="utf-8")) or {}
-            lcr = rules.get("low_confidence_review") or {}
-            threshold = float(lcr.get("threshold", threshold))
-            yaml_cc = list(lcr.get("cc") or [])
-            if yaml_cc:
-                extra_cc = yaml_cc  # YAML 覆蓋預設
-    except Exception:
-        pass
-
-    if conf is not None and conf < threshold:
-        meta["require_review"] = True
-        for x in extra_cc:
-            if x not in cc:
-                cc.append(x)
-
-    res["meta"] = meta
-    if cc:
-        res["cc"] = cc
-    return res
-
-
-def apply_policy(result: Dict[str, Any], message: Dict[str, Any], context: Optional[str] = None) -> Dict[str, Any]:
-    """單筆策略代理到 apply_policies。"""
-    return apply_policies(result, message, context or "config/policy.yaml")
+# Compat shim: from src.policy_engine import * → smart_mail_agent.policy_engine
+from smart_mail_agent.policy_engine import *  # noqa: F401,F403
