+++ src/utils/__init__.py
@@ -1,2 +1,1 @@
-# Compat shim: from utils import X → smart_mail_agent.utils.X
 from smart_mail_agent.utils import *  # noqa: F401,F403
