+++ utils/__init__.py
@@ -1,2 +1,2 @@
-# Compat shim at repo root: import utils.* → smart_mail_agent.utils.*
+# Compat shim: import utils.* → smart_mail_agent.utils.*
 from smart_mail_agent.utils import *  # noqa: F401,F403
