+++ utils/mailer.py
@@ -1,1 +1,2 @@
+import smtplib  # keep patchable: tests use patch("utils.mailer.smtplib.SMTP_SSL")
 from smart_mail_agent.utils.mailer import *  # noqa: F401,F403
